Stop Travis-ci building twice
Currently we have travis set to build both PR's and pushes. However this means when we push to a PR we have to wait for Travis to finish 2 builds.

If we unticked 'pushes' when the PR was finally merged that would not be built. The brute force approach would be to untick build PR's and just build all pushes. We instead have gone with the approach outlined here http://stackoverflow.com/a/31882307